Add dialog for confirmation before  delete profile

+++ Hotel.Frontend/src/app/components/personal/personal.component.ts
@@ -1,9 +1,11 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { Router } from '@angular/router';
 import { ReservationListItem } from 'src/app/models/reservation-list-item';
 import { UserModel } from 'src/app/models/user-model';
 import { AccountService } from 'src/app/services/account.service';
 import { ReservationService } from 'src/app/services/reservation.service';
+import { ConfirmationComponent } from '../confirmation/confirmation.component';
 
 @Component({
   selector: 'app-personal',
@@ -15,7 +17,10 @@ export class PersonalComponent implements OnInit {
   currentUser: UserModel | null = null;
   myReservations!: Array<ReservationListItem>; 
 
-  constructor(private accountService: AccountService, private reservationService: ReservationService, private router: Router){}
+  constructor(private accountService: AccountService,
+              private reservationService: ReservationService,
+              private router: Router,
+              private dialog: MatDialog){}
   
   ngOnInit(): void {
     this.accountService.user.subscribe({
@@ -33,8 +38,27 @@ export class PersonalComponent implements OnInit {
   }
 
   async deleteProfile(userId: string){
-    await this.accountService.deleteProfile(userId)
-    .then(() => this.router.navigate(['']))
+
+    let dialogBoxSettings = {
+      width: '400px',
+      margin: '0 auto',
+      disableClose: true,
+      hasBackdrop: true,
+      position: {top: '10%'},
+      data: {message: "Biztosan törlöd a profilodat?"}
+    };
+
+    let dialogRef = this.dialog.open(ConfirmationComponent, dialogBoxSettings);
+    dialogRef.afterClosed().subscribe({
+      next: (res) => {
+        if(res === "agree"){
+          this.accountService.deleteProfile(userId)
+            .then(() => this.router.navigate(['']))
+        }
+      }
+    })
+
+    
   }
 
   refreshReservations(message: string){
